@@ -1,14 +1,21 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { BrowserRouter } from 'react-router-dom'
 
-ReactDOM.render(
+
+const domNode = document.getElementById('root');
+const root = createRoot(domNode);
+
+root.render(
   <React.StrictMode>
+    <BrowserRouter>
     <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+    </BrowserRouter>
+  </React.StrictMode>
+
 );
 
 // If you want to start measuring performance in your app, pass a function
