@@ -1,6 +1,6 @@
 class Review < ApplicationRecord
-    has_many :movies
-    has_many :lists, through: :movies
+   belongs_to :user
+   belongs_to :movie
 
 
 end
