@@ -1,4 +1,5 @@
 class User < ApplicationRecord
-    has_many :lists
+    has_many :reviews
+    has_many :movies, through: :reviews
     has_secure_password
 end
