@@ -1,20 +1,22 @@
 import React from 'react'
-import { Link } from "react-router-dom"
+import { useNavigate } from "react-router-dom"
 
 function Home() {
+    const navigate = useNavigate();
 
     // I want the Login form to dissapear if user is logged in, and
     // I want it to say Welcome <User>! 
   return (
     <div>
-        Have User login Here
+      <h2>Login: </h2>
         <form>
             <label>Username:</label>
             <input name="username" type="text" ></input>
             <label>Password:</label>
             <input name="username" type="text" ></input>
+            <input type="submit" />
         </form>
-        Have a button to make an account
+        <button onClick={() => navigate(`users/new`)}>Create an Account</button>
     </div>
   )
 }
