@@ -10,16 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2023_07_11_191151) do
-
-  create_table "lists", force: :cascade do |t|
-    t.string "title"
-    t.string "description"
-    t.string "image_url"
-    t.integer "user_id"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
-  end
+ActiveRecord::Schema.define(version: 2023_07_10_220956) do
 
   create_table "movies", force: :cascade do |t|
     t.string "title"
@@ -38,6 +29,8 @@ ActiveRecord::Schema.define(version: 2023_07_11_191151) do
     t.string "review"
     t.boolean "watched"
     t.integer "rating"
+    t.integer "user_id"
+    t.integer "movie_id"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
   end
